First ESP8266 port, and more!
1. Simba runs in a FreeRTOS thread.

The kernel supports everything except:

    - the system tick is hard coded.
    - FSTR() has no effect. Strings will be stored in RAM.

Driver support:

    - pin
    - uart

2. Use scripts to generate parts of the documentation.

+++ src/kernel/std.c
@@ -181,7 +181,7 @@ static char *formatf(char c,
     value = va_arg(*ap_p, double);
 
     /* Convert a negative value a to positive. */
-    if (value < 0.0) {
+    if ((int)value < 0) {
         value *= -1.0;
         *negative_sign_p = 1;
     }
